+++ src/components/CampaignPage/CampaignPage.css
@@ -1,0 +1,4 @@
+.campaign-page-container {
+    display: flex;
+    flex-direction: column;
+}
